feat: cross-app communication
Apps can now call `world.emit(name, data)` to emit world events for other apps to listen in on with `world.on(name, callback)`.

This allows cross app communication!

NOTE: we dont sanitize `data` emitted but its best to keep this as a plain old javascript object and not send things like functions to avoid issues.