fix: tiny octree insertions
Tiny objects and zero radius objects being added to the octree were causing the octree to build a large number of subdivisions. This was causing stutters in performance and in some cases straight up freezing.

Commit adds a minimum radius for all objects in the octree, to prevent subdivisions going so deep.